model/code/requirements.txt
astroid==2.4.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
boto3==1.13.19
botocore==1.16.19
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
protobuf3-to-dict==0.1.5
pycodestyle==2.6.0
Pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.1
requests==2.23.0
s3transfer==0.3.3
sagemaker==1.60.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
smdebug-rulesconfig==0.1.2
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
toml==0.10.1
torch==1.5.0+cpu
torchvision==0.6.0+cpu
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.12.1
zipp==3.1.0
